+++ com.unity.cinemachine/Editor/Editors/CinemachineHardLookAtEditor.cs
@@ -1,4 +1,5 @@
 using UnityEditor;
+using UnityEditor.UIElements;
 using UnityEngine.UIElements;
 
 namespace Unity.Cinemachine.Editor
@@ -7,10 +8,13 @@ namespace Unity.Cinemachine.Editor
     [CanEditMultipleObjects]
     class CinemachineHardLookAtEditor : UnityEditor.Editor
     {
+        CinemachineHardLookAt Target => target as CinemachineHardLookAt;
+
         public override VisualElement CreateInspectorGUI()
         {
             var ux = new VisualElement();
             this.AddMissingCmCameraHelpBox(ux, CmPipelineComponentInspectorUtility.RequiredTargets.LookAt);
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.LookAtOffset)));
             return ux;
         }
     }