+++ com.unity.cinemachine/Editor/Obsolete/CinemachineVirtualCameraEditor.cs
@@ -22,7 +22,7 @@ namespace Unity.Cinemachine.Editor
             var brain = CinemachineCore.FindPotentialTargetBrain(vcam);
             if (brain != null)
             {
-                vcam.Lens = brain.State.Lens;
+                vcam.m_Lens.SetFromLensSettings(brain.State.Lens);
                 vcam.transform.SetPositionAndRotation(brain.transform.position, brain.transform.rotation);
             }
         }
@@ -31,7 +31,7 @@ namespace Unity.Cinemachine.Editor
         static void AdoptSceneViewCameraSettings(MenuCommand command)
         {
             var vcam = command.context as CinemachineVirtualCamera;
-            vcam.Lens = CinemachineMenu.MatchSceneViewCamera(vcam.transform);
+            vcam.m_Lens.SetFromLensSettings(CinemachineMenu.MatchSceneViewCamera(vcam.transform));
         }
         
         /// <summary>Get the property names to exclude in the inspector.  
@@ -125,13 +125,13 @@ namespace Unity.Cinemachine.Editor
             if (CinemachineSceneToolUtility.IsToolActive(typeof(FoVTool)))
             {
                 CinemachineSceneToolHelpers.FovToolHandle(vcam, 
-                    new SerializedObject(vcam).FindProperty(() => vcam.Lens), 
-                    vcam.Lens, IsHorizontalFOVUsed());
+                    new SerializedObject(vcam).FindProperty(() => vcam.m_Lens), 
+                    vcam.m_Lens.ToLensSettings(), IsHorizontalFOVUsed());
             }
             else if (CinemachineSceneToolUtility.IsToolActive(typeof(FarNearClipTool)))
             {
                 CinemachineSceneToolHelpers.NearFarClipHandle(vcam, 
-                    new SerializedObject(vcam).FindProperty(() => vcam.Lens));
+                    new SerializedObject(vcam).FindProperty(() => vcam.m_Lens));
             }
             Handles.color = originalColor;
         }
@@ -153,7 +153,7 @@ namespace Unity.Cinemachine.Editor
             DrawPropertyInInspector(serializedObject.FindProperty(() => Target.OutputChannel));
             DrawTargetsInInspector(serializedObject.FindProperty(() => Target.m_Follow), serializedObject.FindProperty(() => Target.m_LookAt));
             DrawPropertyInInspector(serializedObject.FindProperty(() => Target.StandbyUpdate));
-            DrawPropertyInInspector(serializedObject.FindProperty(() => Target.Lens));
+            DrawPropertyInInspector(serializedObject.FindProperty(() => Target.m_Lens));
             DrawRemainingPropertiesInInspector();
             m_PipelineSet.OnInspectorGUI(!IsPropertyExcluded("Header"));
             DrawNonExcludedExtensionsWidgetInInspector();