+++ com.unity.cinemachine/CHANGELOG.md
@@ -15,10 +15,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added CinemachineVirtualCameraBase.CancelDamping() convenience method to snap camera to its target position.
 - Added CinemachineOrbitalFollow.TargetOffset to reposition orbit center.
-- SaveDuringPlay supports multi-scene editing.
+- Added CinemachineGroupFraming.CenterOffset to reposition group center on the screen.
+- Added LookAtOffset to CinemachineHardLookAt behaviour.
 
 ### Changed
 - RuntimeUtility.GetScratchCollider and RuntimeUtility.DestroyScratchCollider are now public, to allow custom extensions to use them.
+- SaveDuringPlay supports multi-scene editing.
 
 
 ## [3.0.0] - 2023-10-25